feat: allow providing custom runner per transform file thru CodeshiftConfig
depends on:
- https://github.com/CodeshiftCommunity/CodeshiftCommunity/pull/63

example on how it could be used:
- https://github.com/pipedrive/CodeshiftCommunity/pull/22
    - though note we might refactor into separate PRs, idk,
      preferably would use directly from upstream (you).

fixes a lot of cases for us:

1. we have a postcss codemod that we want to run,
while still utilizing the @codeshift/cli.

though, i don't know if these changes will work
if we're using a remote package, will they?

2. we'll want to do some global pre-processing
on files before running our codemod.

though, there's still no way to provide the codemod
as a __function__ instead of an __import path__ to jscodeshift,
which will force us to do dependency injection
instead of just passing the pre-processed results
as an argument to a function.

this is where the considerations to fork jscodeshift
come into play again:
- https://github.com/CodeshiftCommunity/CodeshiftCommunity/issues/67

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>